backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
pydantic==2.10.3
python-dotenv==1.0.1
aiofiles==24.1.0
google-genai==1.5.0
PyJWT==2.10.1
python-multipart==0.0.20
asyncpg==0.30.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
